# GH-98686: Get rid of `BINARY_OP_GENERIC` and `COMPARE_OP_GENERIC`