--- a/Lib/opcode.py
+++ b/Lib/opcode.py
@@ -282,7 +282,6 @@ _specializations = {
         "BINARY_OP_ADD_FLOAT",
         "BINARY_OP_ADD_INT",
         "BINARY_OP_ADD_UNICODE",
-        "BINARY_OP_GENERIC",
         "BINARY_OP_INPLACE_ADD_UNICODE",
         "BINARY_OP_MULTIPLY_FLOAT",
         "BINARY_OP_MULTIPLY_INT",
@@ -316,7 +315,6 @@ _specializations = {
     ],
     "COMPARE_OP": [
         "COMPARE_OP_FLOAT_JUMP",
-        "COMPARE_OP_GENERIC",
         "COMPARE_OP_INT_JUMP",
         "COMPARE_OP_STR_JUMP",
     ],