--- a/Misc/NEWS.d/next/Core and Builtins/2022-11-10-17-09-16.gh-issue-98686.bmAKwr.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2022-11-10-17-09-16.gh-issue-98686.bmAKwr.rst
@@ -1,0 +1,2 @@
+Remove the ``BINARY_OP_GENERIC`` and ``COMPARE_OP_GENERIC``
+"specializations".