… Include/internal/pycore_opcode.h
… Include/opcode.h
… Python/opcode_targets.h