@brandtbucher 2022-11-11
> Thanks for the merge conflict. :-/

I'll wait until after #99313 is merged to land this. Anything else in flight that I should wait for?

> Don't you need to change the magic number? All the opcodes are different.

We don't need to bump the magic number when adding/removing/changing specialized instructions or superinstructions, since those never get included in pycs! It's a pretty nice design.