@gvanrossum 2022-11-13
In that case I welcome your review of #99313.